service/air_particles/requirements.txt
influxdb==3.0.0
pyserial==3.4
